requirements.txt
google-generativeai
python-dotenv
litellm
langchain-community
crewai
